JobOutput: extract JobOutputSearch bar
